charts: move chart to folder in containers, support routing
